@@ -53,7 +53,7 @@ class SendtoallCommand extends AdminCommand
      * @return ServerResponse
      * @throws TelegramException
      */
-    public function execute()
+    public function execute(): ServerResponse
     {
         $text = $this->getMessage()->getText(true);
 
