@@ -45,7 +45,7 @@ class StartCommand extends UserCommand
      *
      * @return ServerResponse
      */
-    public function execute()
+    public function execute(): ServerResponse
     {
         //$message = $this->getMessage();
         //$chat_id = $message->getChat()->getId();
