@@ -41,7 +41,7 @@ class ChatMember extends Entity
     /**
      * {@inheritdoc}
      */
-    protected function subEntities()
+    protected function subEntities(): array
     {
         return [
             'user' => User::class,
