@@ -35,7 +35,7 @@ class Game extends Entity
     /**
      * {@inheritdoc}
      */
-    protected function subEntities()
+    protected function subEntities(): array
     {
         return [
             'photo'         => [PhotoSize::class],
