@@ -30,7 +30,7 @@ class GameHighScore extends Entity
     /**
      * {@inheritdoc}
      */
-    protected function subEntities()
+    protected function subEntities(): array
     {
         return [
             'user' => User::class,
