@@ -24,7 +24,7 @@ class ReplyToMessage extends Message
      * @param array  $data
      * @param string $bot_username
      */
-    public function __construct(array $data, $bot_username = '')
+    public function __construct(array $data, string $bot_username = '')
     {
         //As explained in the documentation
         //Reply to message can't contain other reply to message entities
