@@ -32,7 +32,7 @@ class Sticker extends Entity
     /**
      * {@inheritdoc}
      */
-    protected function subEntities()
+    protected function subEntities(): array
     {
         return [
             'thumb'         => PhotoSize::class,
