@@ -28,7 +28,7 @@ class StickerSet extends Entity
     /**
      * {@inheritdoc}
      */
-    protected function subEntities()
+    protected function subEntities(): array
     {
         return [
             'stickers' => [Sticker::class],
