@@ -28,7 +28,7 @@ class PassportData extends Entity
     /**
      * {@inheritdoc}
      */
-    protected function subEntities()
+    protected function subEntities(): array
     {
         return [
             'data'        => [EncryptedPassportElement::class],
